requirements.txt
flask==3.1.0
flask-cors==4.0.0
psycopg2==2.9.9
requests==2.32.5
python-dotenv==1.0.1
